Change from address for inquiry emails
- Previously, the user's entered email address was being used as the
"from" address for emails sent via the contact form. That works okay,
but there's a risk that the Census email account it's sent to might put
them in the spam folder. If there's a constant, known from address, the
Census Department can whitelist that address so that the emails always
get through. It would also allow them to set up a filter or rule to
apply to all emails from the contact form (with the individual user's
email as the from address, they could only do that by matching on the
subject line or something, which is much more fragile). The user's email
is now set as the reply-to address, so the Census Department will still
be able to reply to the requester simply by hitting "Reply".